Merged PR 125: Fix wrong Context casting
This casting is not needed since in the SurfaceDuoLayoutStatusHandler we expect now a Context and not an Activity.
This was also crashing apps that were using aonther kind of Context as ContextThemeWrapper which Activity doesn't inherit from it.

This branch/PR is meant **to be merged into** **release/alpha02** branch which is from where it was branched originally.